# Improve sketching performance for DNA