In #860 I described a performance issue in the sketching of DNA sequences: characters were repeatedly rechecked for their validity. The approach by @luizirber vastly speeds up the checking of characters (#861). However, most characters were still checked up to k times. Also, some valid k-mers were now missed due to a small logic error.

This pull request comes in three parts.

1. An initially failing unit test, showcasing the regression introduced in a07dd85.
2. Avoid rechecking valid characters.
3. Remove fast path as it was slowing us down.

Here are the benchmarks:

```
 add_sequence/valid      time:   [4.9079 ms 4.9765 ms 5.0571 ms]                               
                        change: [-7.5285% -6.2820% -4.8922%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking add_sequence/lowercase: Collecting 10 samples in estimated 5.1661 s (1045 iterati                                                                                              add_sequence/lowercase  time:   [4.9820 ms 5.1157 ms 5.3481 ms]
                        change: [-9.3013% -6.2065% -2.4192%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking add_sequence/invalid kmers: Collecting 10 samples in estimated 5.1394 s (1540 ite                                                                                              add_sequence/invalid kmers                        
                        time:   [3.2999 ms 3.3375 ms 3.3611 ms]
                        change: [-32.280% -31.194% -29.966%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
Benchmarking add_sequence/force with valid kmers: Collecting 10 samples in estimated 5.1789 s                                                                                               add_sequence/force with valid kmers                        
                        time:   [4.8650 ms 4.9078 ms 4.9668 ms]
                        change: [-11.240% -7.9411% -4.2589%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
  1 (10.00%) high severe
```

This code is not particularly Rusty as I don't know the idioms. Feel free to request changes. Also I don't have all python dependencies installed, so please run the checks again.

Best,
Fabian

- [x] Is it mergeable?
- [ ] `make test` Did it pass the tests?
- [ ] `make coverage` Is the new code covered?
- [x] Did it change the command-line interface? Only additions are allowed
  without a major version increment. Changing file formats also requires a
  major version number increment.
- [ ] Was a spellchecker run on the source code and documentation after
  changes were made?