@kloetzl 2020-01-26
> (and thanks for keeping the comment, took me some time to draw it in ASCII :sweat_smile: )

At some point during the refactoring one of the unit tests failed, and I was like *what, why?*. I had to consult the diagram to see I got the boundaries wrong.