@kloetzl 2020-01-26
Seeing as you already merged the PR, it might be easiest if you would create a commit with the recommended changes yourself.