@luizirber 2020-01-26
> ~Seeing as you already merged the PR, it might be easiest if you would create a commit with the recommended changes yourself.~ And you were quicker than I could type.

Since #856 is about to be merged too I thought it was easier to fix it there =]

> Thanks for merging, btw. +1

Oh, BTW: I'm adding contributors in #837, so if you're OK with being an author in the sourmash 3.0 paper can you paste your ORCID ID here (so I can add it there)?