@kloetzl 2020-01-26
> Oh, BTW: I'm adding contributors in #837, so if you're OK with being an author in the sourmash 3.0 paper can you paste your ORCID ID here (so I can add it there)?

Sure, I would be happy about that. But I don't have an ORCID id.