@ctb 2020-01-26
Thank you for all your help! You can get one at https://orcid.org/register - I think our journal of choice (JOSS) requires them, unfortunately :(