@kloetzl 2020-01-27
Done: https://orcid.org/0000-0002-6930-0592